+++ push/model/src/main/java/org/jboss/aerogear/unifiedpush/message/apns/APNs.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.unifiedpush.message.ios;
+package org.jboss.aerogear.unifiedpush.message.apns;
 
 import org.codehaus.jackson.annotate.JsonProperty;
 import org.codehaus.jackson.map.annotate.JsonRootName;
